backend/requirements.txt
fastapi
uvicorn[standard]
kiteconnect
xgboost
shap
openai
python-dotenv
httpx
pydantic>=2.0
